Move CSV::Table to a separate file
+ Moved the CSV::Table class to a separate file

This PR Implements part of #5.

It does not change any behavior of the csv gem.
